Refactor reporting feature components to a feature-based approach
