+++ frontend/src/api/APIWorker.tsx
@@ -12,11 +12,11 @@ import { openBeaconMalfunctionInKanban } from '../domain/use_cases/beaconMalfunc
 import getAllGearCodes from '../domain/use_cases/gearCode/getAllGearCodes'
 import { getInfractions } from '../domain/use_cases/infraction/getInfractions'
 import { getVesselControls } from '../domain/use_cases/mission/getVesselControls'
-import { getAllCurrentReportings } from '../domain/use_cases/reporting/getAllCurrentReportings'
 import getAllSpecies from '../domain/use_cases/species/getAllSpecies'
 import { getVesselReportings } from '../domain/use_cases/vessel/getVesselReportings'
 import { updateVesselTracks } from '../domain/use_cases/vessel/updateVesselTracks'
 import getAllRegulatoryLayers from '../features/Regulation/useCases/getAllRegulatoryLayers'
+import { getAllCurrentReportings } from '../features/Reporting/useCases/getAllCurrentReportings'
 import { useIsSuperUser } from '../hooks/authorization/useIsSuperUser'
 import { useMainAppDispatch } from '../hooks/useMainAppDispatch'
 import { useMainAppSelector } from '../hooks/useMainAppSelector'
