+++ frontend/src/domain/shared_slices/Vessel.ts
@@ -1,8 +1,8 @@
 import { createSlice } from '@reduxjs/toolkit'
 import { transform } from 'ol/proj'
 
+import { reportingIsAnInfractionSuspicion, ReportingTypeCharacteristics } from '../../features/Reporting/types'
 import { OPENLAYERS_PROJECTION, WSG84_PROJECTION } from '../entities/map/constants'
-import { reportingIsAnInfractionSuspicion, ReportingTypeCharacteristics } from '../entities/reporting'
 import {
   atLeastOneVesselSelected,
   getOnlyVesselIdentityProperties,
