+++ frontend/src/domain/use_cases/vessel/unselectVessel.ts
@@ -1,7 +1,7 @@
 import { logbookActions } from '../../../features/Logbook/slice'
+import { resetCurrentAndArchivedReportingsOfSelectedVessel } from '../../../features/Reporting/slice'
 import { resetVesselBeaconMalfunctionsResumeAndHistory } from '../../shared_slices/BeaconMalfunction'
 import { expandRightMenu } from '../../shared_slices/Global'
-import { resetCurrentAndArchivedReportingsOfSelectedVessel } from '../../shared_slices/Reporting'
 import { closeVesselSidebar, resetSelectedVessel } from '../../shared_slices/Vessel'
 
 export const unselectVessel = () => dispatch => {
