+++ frontend/src/features/Reporting/components/ReportingList/index.tsx
@@ -8,28 +8,28 @@ import * as timeago from 'timeago.js'
 
 import { REPORTING_LIST_TABLE_OPTIONS } from './constants'
 import { getReportingOrigin, getReportingTitle } from './utils'
-import { ALERTS_MENU_SEA_FRONT_TO_SEA_FRONTS } from '../../../../../domain/entities/alerts/constants'
-import { SeaFrontGroup } from '../../../../../domain/entities/seaFront/constants'
-import { setEditedReportingInSideWindow } from '../../../../../domain/shared_slices/Reporting'
-import { ReportingType } from '../../../../../domain/types/reporting'
-import archiveReportings from '../../../../../domain/use_cases/reporting/archiveReportings'
-import deleteReportings from '../../../../../domain/use_cases/reporting/deleteReportings'
-import { showVessel } from '../../../../../domain/use_cases/vessel/showVessel'
-import { useForceUpdate } from '../../../../../hooks/useForceUpdate'
-import { useMainAppDispatch } from '../../../../../hooks/useMainAppDispatch'
-import { useMainAppSelector } from '../../../../../hooks/useMainAppSelector'
-import { useTable } from '../../../../../hooks/useTable'
-import { CardTable } from '../../../../../ui/card-table/CardTable'
-import { CardTableBody } from '../../../../../ui/card-table/CardTableBody'
-import { CardTableFilters } from '../../../../../ui/card-table/CardTableFilters'
-import { CardTableRow } from '../../../../../ui/card-table/CardTableRow'
-import { EmptyCardTable } from '../../../../../ui/card-table/EmptyCardTable'
-import { FilterTableInput } from '../../../../../ui/card-table/FilterTableInput'
-import { downloadAsCsv } from '../../../../../utils/downloadAsCsv'
-import { Flag } from '../../../../VesselList/tableCells'
-import { EditReporting } from '../EditReporting'
+import { ALERTS_MENU_SEA_FRONT_TO_SEA_FRONTS } from '../../../../domain/entities/alerts/constants'
+import { SeaFrontGroup } from '../../../../domain/entities/seaFront/constants'
+import { ReportingType } from '../../../../domain/types/reporting'
+import { showVessel } from '../../../../domain/use_cases/vessel/showVessel'
+import { useForceUpdate } from '../../../../hooks/useForceUpdate'
+import { useMainAppDispatch } from '../../../../hooks/useMainAppDispatch'
+import { useMainAppSelector } from '../../../../hooks/useMainAppSelector'
+import { useTable } from '../../../../hooks/useTable'
+import { CardTable } from '../../../../ui/card-table/CardTable'
+import { CardTableBody } from '../../../../ui/card-table/CardTableBody'
+import { CardTableFilters } from '../../../../ui/card-table/CardTableFilters'
+import { CardTableRow } from '../../../../ui/card-table/CardTableRow'
+import { EmptyCardTable } from '../../../../ui/card-table/EmptyCardTable'
+import { FilterTableInput } from '../../../../ui/card-table/FilterTableInput'
+import { downloadAsCsv } from '../../../../utils/downloadAsCsv'
+import { EditReporting } from '../../../SideWindow/Alert/AlertListAndReportingList/EditReporting'
+import { Flag } from '../../../VesselList/tableCells'
+import { setEditedReportingInSideWindow } from '../../slice'
+import archiveReportings from '../../useCases/archiveReportings'
+import deleteReportings from '../../useCases/deleteReportings'
 
-import type { InfractionSuspicionReporting, PendingAlertReporting } from '../../../../../domain/types/reporting'
+import type { InfractionSuspicionReporting, PendingAlertReporting } from '../../../../domain/types/reporting'
 import type { CSSProperties, MutableRefObject } from 'react'
 
 type ReportingListProps = {
@@ -38,7 +38,7 @@ type ReportingListProps = {
 export function ReportingList({ selectedSeaFrontGroup }: ReportingListProps) {
   const dispatch = useMainAppDispatch()
   const searchInputRef = useRef() as MutableRefObject<HTMLInputElement>
-  const { currentReportings } = useMainAppSelector(state => state.reporting)
+  const currentReportings = useMainAppSelector(state => state.reporting.currentReportings)
   const { forceDebouncedUpdate } = useForceUpdate()
 
   const baseUrl = useMemo(() => window.location.origin, [])
