+++ frontend/src/features/Reporting/components/VesselReportings/Current/ConfirmDeletionModal.tsx
@@ -1,9 +1,9 @@
 import { Modal } from 'rsuite'
 import styled from 'styled-components'
 
-import { COLORS } from '../../../../constants/constants'
-import StyledModalHeader from '../../../commonComponents/StyledModalHeader'
-import { CancelButton, ValidateButton } from '../../../commonStyles/Buttons.style'
+import { COLORS } from '../../../../../constants/constants'
+import StyledModalHeader from '../../../../commonComponents/StyledModalHeader'
+import { CancelButton, ValidateButton } from '../../../../commonStyles/Buttons.style'
 
 export function ConfirmDeletionModal({ closeModal, isOpened, validateCallback }) {
   return (
