+++ frontend/src/features/Reporting/types.ts
@@ -1,6 +1,6 @@
-import { ReportingType } from '../types/reporting'
+import { ReportingType } from '../../domain/types/reporting'
 
-import type { Reporting } from '../types/reporting'
+import type { Reporting } from '../../domain/types/reporting'
 
 type ReportingTypeCharacteristic = {
   // TODO It should be useless now that types are discriminated.
