+++ frontend/src/features/SideWindow/index.tsx
@@ -19,17 +19,17 @@ import { MissionForm } from './MissionForm'
 import { MissionList } from './MissionList'
 import { SideWindowMenuKey } from '../../domain/entities/sideWindow/constants'
 import { closeBeaconMalfunctionInKanban } from '../../domain/shared_slices/BeaconMalfunction'
-import { setEditedReportingInSideWindow } from '../../domain/shared_slices/Reporting'
 import { getOperationalAlerts } from '../../domain/use_cases/alert/getOperationalAlerts'
 import { getSilencedAlerts } from '../../domain/use_cases/alert/getSilencedAlerts'
 import getAllBeaconMalfunctions from '../../domain/use_cases/beaconMalfunction/getAllBeaconMalfunctions'
 import getAllGearCodes from '../../domain/use_cases/gearCode/getAllGearCodes'
 import { getInfractions } from '../../domain/use_cases/infraction/getInfractions'
-import { getAllCurrentReportings } from '../../domain/use_cases/reporting/getAllCurrentReportings'
 import { sideWindowDispatchers } from '../../domain/use_cases/sideWindow'
 import { useMainAppDispatch } from '../../hooks/useMainAppDispatch'
 import { useMainAppSelector } from '../../hooks/useMainAppSelector'
 import { FrontendErrorBoundary } from '../../ui/FrontendErrorBoundary'
+import { setEditedReportingInSideWindow } from '../Reporting/slice'
+import { getAllCurrentReportings } from '../Reporting/useCases/getAllCurrentReportings'
 
 export type SideWindowProps = HTMLAttributes<HTMLDivElement> & {
   isFromURL: boolean
