+++ frontend/src/store/reducers.ts
@@ -19,7 +19,6 @@ import layer from '../domain/shared_slices/Layer'
 import { mapReducer } from '../domain/shared_slices/Map'
 import { measurementReducer } from '../domain/shared_slices/Measurement'
 import { missionReducer, type MissionState } from '../domain/shared_slices/Mission'
-import { reportingReducer } from '../domain/shared_slices/Reporting'
 import { sideWindowReducer } from '../domain/shared_slices/SideWindow'
 import { speciesReducer } from '../domain/shared_slices/Species'
 import { vesselSliceReducer } from '../domain/shared_slices/Vessel'
@@ -30,6 +29,7 @@ import { customZoneReducer, type CustomZoneState } from '../features/CustomZone/
 import { logbookReducer } from '../features/Logbook/slice'
 import { regulatoryLayerSearchReducer } from '../features/Regulation/components/RegulatoryZones/search/slice'
 import { regulatoryReducer } from '../features/Regulation/slice'
+import { reportingReducer } from '../features/Reporting/slice'
 import { alertReducer } from '../features/SideWindow/Alert/slice'
 import { stationReducer } from '../features/Station/slice'
 import { vesselListReducer } from '../features/VesselList/slice'
